+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 * Added a changelog
 * Switched to using [httpx](https://www.python-httpx.org/) for HTTP requests instead of [requests](https://requests.readthedocs.io/en/master/)
 * Allow overriding the default HTTP client with a custom one
+* Add SDK methods for managing retention durations for soft-deleted data